+++ dropwizard-core/src/main/java/com/yammer/dropwizard/json/Json.java
@@ -47,7 +47,7 @@ public class Json {
         mapper.disable(SerializationConfig.Feature.WRITE_ENUMS_USING_TO_STRING);
         mapper.disable(DeserializationConfig.Feature.READ_ENUMS_USING_TO_STRING);
         mapper.registerModule(new GuavaModule());
-        mapper.registerModule(new Log4jModule());
+        mapper.registerModule(new LogbackModule());
 
         this.typeFactory = mapper.getTypeFactory();
     }